requirements.txt
bidict==0.22.0
click==8.1.3
confluent-kafka==1.9.2
Flask==2.1.3
h11==0.13.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-decouple==3.6
python-dotenv==0.20.0
python-engineio==4.3.3
python-socketio==5.7.1
Werkzeug==2.2.1
zipp==3.8.1
